Support for Scratch hosted with https
I've modified the URLs for submitting twitter API calls to remove the explicit 'http' scheme. I've done this to avoid Mixed Content errors from browsers that will block the request when the parent Scratch page is being hosted at an 'https' address. This should not prevent the current behaviour of working from http.